internal/c2go: finalize int32 conversion for consistent overflow behavior
Complete the int32 conversion across all c2go functions and tests to ensure
consistent overflow behavior with H3 C implementation. Updates function
signatures, test parameters, and type casting to use int32 consistently.

+++ internal/c2go/baseCells___baseCellIsCwOffset.go
@@ -3,7 +3,7 @@ package c2go
 // _baseCellIsCwOffset returns true if the given face is a clockwise offset face for the given pentagon base cell.
 // For non-pentagon base cells, this always returns false.
 // Ported from H3 C: baseCells.c::_baseCellIsCwOffset
-func _baseCellIsCwOffset(baseCell int, testFace int) bool {
-	return baseCellData[baseCell].CwOffsetPent[0] == int32(testFace) ||
-		baseCellData[baseCell].CwOffsetPent[1] == int32(testFace)
+func _baseCellIsCwOffset(baseCell int32, testFace int32) bool {
+	return baseCellData[baseCell].CwOffsetPent[0] == testFace ||
+		baseCellData[baseCell].CwOffsetPent[1] == testFace
 }
