+++ internal/c2go/baseCells___baseCellToCCWrot60.go
@@ -4,7 +4,7 @@ package c2go
 // a given face using the faceIjkBaseCells lookup table.
 // Returns INVALID_ROTATIONS if the base cell is not found on the face.
 // Ported from H3 C: baseCells.c::_baseCellToCCWrot60
-func _baseCellToCCWrot60(baseCell int, face int) int {
+func _baseCellToCCWrot60(baseCell int32, face int32) int32 {
 	if face < 0 || face > NUM_ICOSA_FACES {
 		return INVALID_ROTATIONS
 	}
@@ -19,8 +19,8 @@ func _baseCellToCCWrot60(baseCell int, face int) int {
 	for i := 0; i < 3; i++ {
 		for j := 0; j < 3; j++ {
 			for k := 0; k < 3; k++ {
-				if faceIjkBaseCells[face][i][j][k].BaseCell == int32(baseCell) {
-					return int(faceIjkBaseCells[face][i][j][k].CcwRot60)
+				if faceIjkBaseCells[face][i][j][k].BaseCell == baseCell {
+					return faceIjkBaseCells[face][i][j][k].CcwRot60
 				}
 			}
 		}
