+++ internal/c2go/baseCells__getRes0Cells_parity_test.go
@@ -38,13 +38,13 @@ func TestGetRes0CellsInvalidSize(t *testing.T) {
 
 func TestSetBaseCellParity(t *testing.T) {
 	// Test setBaseCell helper function
-	for bc := 0; bc < NUM_BASE_CELLS; bc++ {
+	for bc := int32(0); bc < NUM_BASE_CELLS; bc++ {
 		h := H3Index(H3_INIT)
 		h = setMode(h, H3_CELL_MODE)
 		h = H3Index(setBaseCell(uint64(h), bc))
 
 		// Verify the base cell was set correctly by extracting it
-		extractedBC := int((uint64(h) & H3_BC_MASK) >> H3_BC_OFFSET)
+		extractedBC := int32((uint64(h) & H3_BC_MASK) >> H3_BC_OFFSET)
 		if extractedBC != bc {
 			t.Errorf("Base cell %d: expected %d, got %d", bc, bc, extractedBC)
 		}
