+++ internal/c2go/h3index___h3ToFaceIjkWithInitializedFijk.go
@@ -4,19 +4,19 @@ package c2go
 // The FaceIJK address should be initialized with the desired face and normalized base cell coordinates.
 // Returns 1 if the possibility of overage exists, otherwise 0.
 // Ported from H3 C: h3Index.c::_h3ToFaceIjkWithInitializedFijk
-func _h3ToFaceIjkWithInitializedFijk(h H3Index, fijk *FaceIJK) int {
+func _h3ToFaceIjkWithInitializedFijk(h H3Index, fijk *FaceIJK) int32 {
 	ijk := &fijk.Coord
 	res := getResolution(h)
 
 	// Center base cell hierarchy is entirely on this face
-	possibleOverage := 1
+	possibleOverage := int32(1)
 	if !_isBaseCellPentagon(getBaseCellNumber(h)) &&
 		(res == 0 ||
 			(fijk.Coord.I == 0 && fijk.Coord.J == 0 && fijk.Coord.K == 0)) {
 		possibleOverage = 0
 	}
 
-	for r := 1; r <= res; r++ {
+	for r := int32(1); r <= res; r++ {
 		if isResolutionClassIII(r) {
 			// Class III == rotate ccw
 			_downAp7(ijk)
