+++ internal/c2go/h3index__hasChildAtRes_parity_test.go
@@ -7,7 +7,7 @@ import "testing"
 func Test_h3index_hasChildAtRes_ParityWithC(t *testing.T) {
 	hs := []H3Index{0x8928308280fffff, 0x821c07fffffffff}
 	for _, h := range hs {
-		for childRes := -1; childRes <= 16; childRes++ {
+		for childRes := int32(-1); childRes <= 16; childRes++ {
 			goVal := _hasChildAtRes(h, childRes)
 			cVal := hasChildAtResC(h, childRes) != 0
 			if goVal != cVal {
